Some fixes to the layout.
Display the navbar in DDDEA green (i.e., don't use inverse colouring),
and try to get it to collapse correctly, with the login drop-down next
to the three-bar button for the menu.  This mostly works, but when
resized down the navbar annoyingly jumps down about 50px; haven't worked
out why yet.