Add an HTML Helper for formatting the event name.
The text string "DDD East Anglia" appears in a few places in the site,
formatted in a specific way, so it seems to make sense to pull this out
into an HTML Helper.  Now we can do @Html.DDDEastAnglia() and get the
correct formatting without having to remember all the time.