Add an image_src link.
Apparently this works with embed.ly to enable Twitter Cards for our site
when someone posts a link to our site using their service. @adrianbanks
suggested it also works with Facebook, so seems like a good change to
make.

Also remove the second favicon link, most likely the result of a
copy-paste error.