Fix the Log In menu links.
The reason this had broken was because there is a form element inside
each li wrapping the a which is actually the link.  This seems to be
done for security reasons (the form included an Html.AntiForgeryToken to
guard against CSRF attacks), but I don't think this is a valid concern
for log on links (or perhaps even the log off link which will suffer the
same styling issue at the moment).

As a result, I've removed the forms from the log in menu.