Fix deployment error.
The previous approach of importing the connection strings from an external
file requires the external file to exist (unlike for the appSettings, it
seems).  Remove this approach and switch to web.config transforms.

The local DB connection string should only be used for Debug builds. As
such, there's no need to place it in an external file, it can just go in
the Web.Debug.config transform file which will overwrite whatever is in
the web.config.