Add SqlPackage to the repository.
This is a handy tool included with SSDT for packaging databases.
Possibly it should be in a different location, but this is where the
other SSDT binaries have gone (on the database-project branch, for
building the database project on TeamCity), and I'm not keen to
duplicate binary files in the repository.