Regenerate the Bootstrap stylesheet.
This is mostly to fix the .btn-primary rendering. I have also incorporated
the hover colour into the Bootstrap stylesheet, so this is removed from
the Site.css stylesheet.

I will document the full set of overrides in #37 so we can refer back to
it if we need to do this again.