Fix #42: Use HTML 5 input types where possible.
Turns out there weren't many places where this could be done, but it's
useful for those fields that are of a particular type (e.g., email, URL,
etc.). The fields are now auto-validated by the browser, and jQuery
validation hooks into the validation API of the browser so error messages
are displayed consistently.