Fix #53: Add new HTML Helper for the Tweet button.
This commit removes the Twitter-styled Tweet button, and replaces it
with a Bootstrap button with a Twitter icon. This resolves the
noticeable delay in loading the Sessions page, and the original issue
reported: the tweet links are (sometimes) including the session url
twice.

The HTML Helper ensures consistent styling of the new tweet button, and
cuts down on duplicated code too.

The encoding and escaping of the tweet text is a little more complex
than I had hoped. I run the text through three separate encoding
methods, as follows:

 - Escape the tweet text with Uri.EscapeDataString() to properly escape
   any characters in the tweet that are URI-unfriendly. The # character
   for hash-tags is an example of this.
 - If a URL is supplied for the tweet, escape the URL using
   Uri.EscapeUriString(). URIs within URIs should be escaped in this way
   to ensure they are properly preserved.
 - HTML-encode the entire tweet to ensure it is safe to render on the
   page. This last step may not be completely necessary, as < and >
   characters will have been escaped in the first step, which should
   be enough to prevent XSS attacks. Seems like a good idea to do
   belt-and-braces, though.