Fix #96: Add a new Layout for the Admin Area.
I've included a start of a new layout for the Admin Area, but it will
need more work in future.

This fixes the immediate issue, but there's a problem with the Create
Session link in the admin area: its using the Admin Route and so 404s.
We can at least access the admin pages again.