Refactor mail sending infrastructure.
`MailMessage` is now a POCO, and so can just be instantiated directly
via its constructor/object initialiser, rather than via a factory. The
tests now test that a specific email is sent (previously, we only tested
that the email was *created*, and never checked that it was actually
delivered). This also requires MailMessage to implement equality so that
NSubstitute can do its stuff.

`IMessenger` notifies a user about particular events on the site that
they are interested in. For example, `SessionCreationMailMessenger`
sends a user the abstract of the talk they have just submitted.
`IMailTemplate` can render a given file to a string for transmission by
an `IMessenger`. `The TokenSubstitutingMailTemplate` reads the file and
replaces the specified tokens with the specified substitutions; the
`MarkdownConvertingMailTemplate` renders the output in Markdown.

`HtmlMailTemplate` wraps the rendered, Markdown-converted plain text
email into an HTML envelope.