Improve the HTML emails.
We don't need a separate HTML version of each email: we can instead
render the plain-text Markdown message template to HTML, and inject that
into the overall HTML email template. This allows us to wrap it up
really nicely so we say "send this email" and everything down to the
`IPostman` implementation receives that mail template; only when the
`IPostman` implementation comes to actually send the email does it get
rendered to HTML, as this is an implementation detail of emailing people.

The HTML email template is embedded into the site's assembly. This
currently results in more logic in `NinjectWebCommon` right now
than I'm happy with, but this can be extracted out into a new class easily
enough.

`IMailTemplate` can now have token substitutions registered against it.
This means we can later remove the substitutions parameter from `Render()`.