Extract a general-purpose EmailMessenger.
This can take an `IMailTemplate` to deliver and an `IPostman` to deliver
it through, and perform the necessary co-ordination to ensure that the
template is appropriately rendered into a `MailMessage` for delivery by
the `IPostman`.

`IMailTemplate`s additionally now support token substitution in the
subject line as well as the message body.