Fix remaining failing Reset Password tests.
These assertions may be too low-level, and it might be better to assert on
a mocked EmailMessengerFactory and created `EmailMessenger`. That starts
to feel a bit like reimplementing the code in Mocks, though, so I've gone
with this approach for now which takes a real `EmailMessengerFactory`,
which in turn takes a mocked `IPostman`, and the assertion remains that
the `IPostman` asked to deliver the expected email message.