Undo PR #334
The changes made broke the password reset emails, meaning that we had to put in a hack so that they are sent in a crappy format.
This undoes the change to put it back to how it was before, with the intention of putting these changes back in without breaking the reset emails.