Ensure that users have confirmed accounts when resetting passwords
Now we've removed the external logins, we have users who don't have confirmed accounts (or even passwords). This ensures that their account is marked as confirmed (as if it isn't the reset password functionality of the membership provider doesn't work), and also sets a random password to ensure that they can't log in until their password has been reset.